--- .env
@@ -1,2 +1,0 @@
-Host="127.0.0.1:8080"
-DatabaseFile="rest-api.db"
