+++ .env.example
@@ -1,0 +1,2 @@
+BIND_ADDRESS="127.0.0.1:8080"
+DATABASE_URL="rest-api.db"
