+++ src/database/models.rs
@@ -2,78 +2,20 @@ use diesel::prelude::*;
 
 use serde::{Deserialize, Serialize};
 
-use super::schema::posts;
-use super::schema::posts::dsl::posts as post_dsl;
-use super::schema::restaurants;
-
-#[derive(Debug, Deserialize, Serialize, Queryable, Insertable)]
-#[table_name = "posts"]
-pub struct Post {
-    pub id: i32,
-    pub title: String,
-    pub body: String,
-    pub published: bool,
-}
+use diesel::dsl::{delete, select, insert_into, exists};
 
 #[derive(Clone, PartialEq, Queryable, Serialize)]
 pub struct Restaurant {
     pub id: i32,
+    pub smiley_restaurant_id: i32,
+    pub name: String,
+    pub address: String,
+    pub zipcode: String,
     pub city: String,
     pub cvr: String,
-    pub longitude: f32,
-    pub latitude: f32,
     pub pnr: String,
-    pub address: String,
-    pub url: String,
-    pub zipcode: String,
-    pub name: String,
-    pub latest_control: Option<i32>,
-    pub second_latest_control: Option<i32>,
-    pub third_latest_control: Option<i32>,
-    pub fourth_latest_control: Option<i32>,
-}
-
-#[derive(Insertable, Deserialize, Serialize)]
-#[table_name = "restaurants"]
-pub struct NewRestaurant {
-    #[serde(alias = "By")]
-    pub city: String,
-
-    #[serde(alias = "cvrnr")]
-    pub cvr: String,
-
-    #[serde(alias = "Geo_Lat")]
     pub latitude: f32,
-
-    #[serde(alias = "Geo_Lng")]
     pub longitude: f32,
-
-    #[serde(alias = "pnr")]
-    pub pnr: String,
-
-    #[serde(alias = "adresse1")]
-    pub address: String,
-
-    #[serde(alias = "URL")]
-    pub url: String,
-
-    #[serde(alias = "postnr")]
-    pub zipcode: String,
-
-    #[serde(alias = "navn1")]
-    pub name: String,
-
-    #[serde(alias = "seneste_kontrol")]
-    pub latest_control: Option<i32>,
-
-    #[serde(alias = "naestseneste_kontrol")]
-    pub second_latest_control: Option<i32>,
-
-    #[serde(alias = "tredjeseneste_kontrol")]
-    pub third_latest_control: Option<i32>,
-
-    #[serde(alias = "fjerdeseneste_kontrol")]
-    pub fourth_latest_control: Option<i32>,
 }
 #[derive(Queryable, Deserialize, Serialize)]
 pub struct Simplerestaurant {
@@ -82,12 +24,12 @@ pub struct Simplerestaurant {
     lng: f32,
 }
 
-use super::schema::restaurants::dsl::restaurants as res_dsl;
+use super::schema::restaurant::dsl::restaurant as res_dsl;
 impl Restaurant {
     pub fn get_all_resturants(conn: &SqliteConnection) -> Vec<Simplerestaurant> {
-        use super::schema::restaurants::dsl::id;
-        use super::schema::restaurants::dsl::latitude;
-        use super::schema::restaurants::dsl::longitude;
+        use super::schema::restaurant::dsl::id;
+        use super::schema::restaurant::dsl::latitude;
+        use super::schema::restaurant::dsl::longitude;
         res_dsl
             .select((id, latitude, longitude))
             .load::<Simplerestaurant>(conn)
@@ -108,8 +50,8 @@ impl Restaurant {
         selng: f32,
         conn: &SqliteConnection,
     ) -> Vec<Restaurant> {
-        use super::schema::restaurants::dsl::latitude;
-        use super::schema::restaurants::dsl::longitude;
+        use super::schema::restaurant::dsl::latitude;
+        use super::schema::restaurant::dsl::longitude;
         res_dsl
             .filter(latitude.lt(nwlat))
             .filter(latitude.gt(selat))
@@ -121,7 +63,7 @@ impl Restaurant {
     }
 
     pub fn search_by_name(query: String, conn: &SqliteConnection) -> Vec<Self> {
-        use super::schema::restaurants::dsl::name;
+        use super::schema::restaurant::dsl::name;
         res_dsl
             .filter(name.like(query + "%"))
             .get_results::<Restaurant>(conn)
@@ -130,7 +72,7 @@ impl Restaurant {
     }
 
     pub fn search_by_zip(query: String, conn: &SqliteConnection) -> Vec<Self> {
-        use super::schema::restaurants::dsl::zipcode;
+        use super::schema::restaurant::dsl::zipcode;
         res_dsl
             .filter(zipcode.eq(query))
             .get_results::<Restaurant>(conn)
@@ -139,7 +81,7 @@ impl Restaurant {
     }
 
     pub fn search_by_city(query: String, conn: &SqliteConnection) -> Vec<Self> {
-        use super::schema::restaurants::dsl::city;
+        use super::schema::restaurant::dsl::city;
         res_dsl
             .filter(city.like(query + "%"))
             .get_results::<Restaurant>(conn)
@@ -148,158 +90,39 @@ impl Restaurant {
     }
 }
 
-impl Post {
-    pub fn list(conn: &SqliteConnection) -> Vec<Self> {
-        post_dsl.load::<Post>(conn).expect("Error loading posts")
-    }
-
-    pub fn by_id(id: i32, conn: &SqliteConnection) -> Option<Self> {
-        post_dsl.find(id).get_result::<Post>(conn).ok()
-    }
-    /*
-    pub fn by_email(email_str: &str, conn: &SqliteConnection) -> Option<Self> {
-        use super::schema::users::dsl::email;
-
-        user_dsl.filter(email.eq(email_str)).first::<Post>(conn).ok()
-    }
-
-    pub fn by_phone(phone_str: &str, conn: &SqliteConnection) -> Option<Self> {
-        use super::schema::users::dsl::phone;
-
-        user_dsl.filter(phone.eq(phone_str)).first::<Post>(conn).ok()
-    }*/
-
-    pub fn create(conn: &SqliteConnection, post: Post) -> Option<Self> {
-        let id = post.id;
-
-        diesel::insert_into(post_dsl)
-            .values(&post)
-            .execute(conn)
-            .expect("Error saving new user");
-
-        Self::by_id(id, conn)
-    }
-
-    /*fn new_user_struct(id: i32, phone: Option<&str>, email: Option<&str>, published: bool) -> Self {
-        Post {
-            id: id.into(),
-            title: title.map(Into::into),
-            body: body.map(Into::into),
-            published: published.into()
-        }
-    }*/
-}
-
-use super::schema::favorites;
-use super::schema::favorites::dsl::favorites as fav_dsl;
+use super::schema::subscription;
 
 #[derive(Debug, Deserialize, Serialize, Queryable, Insertable)]
-#[table_name = "favorites"]
-pub struct Favorites {
+#[table_name = "subscription"]
+pub struct Subscription {
+    pub id: i32,
     pub restaurant_id: i32,
-    pub token_id: String,
+    pub token: String,
 }
-
-impl Favorites {
-    pub fn add_favorite(res_id: i32, u_id: String, conn: &SqliteConnection) -> Self {
-        match Favorites::find_favorite(res_id, u_id.to_string(), conn) {
-            None => {
-                match u_dsl.find(u_id.to_string()).first::<User>(conn).ok() {
-                    None => {
-                        User::new_user(u_id.to_string(), &conn);
-                    }
-                    Some(_) => {}
-                }
-                let new_fav = Favorites {
-                    restaurant_id: res_id,
-                    token_id: u_id.to_string(),
-                };
-                diesel::insert_into(fav_dsl)
-                    .values(&new_fav)
-                    .execute(conn)
-                    .expect("error saving favorite");
-                new_fav
-            }
-            Some(fav) => fav,
-        }
-    }
-
-    pub fn remove_favorite(res_id: i32, u_id: String, conn: &SqliteConnection) {
-        match Favorites::find_favorite(res_id, u_id, conn) {
-            None => {}
-            Some(fav) => {
-                diesel::delete(fav_dsl.find((fav.restaurant_id, fav.token_id)))
-                    .execute(conn)
-                    .expect("error deleting");
+impl Subscription {
+    
+    pub fn subscribe(res_id: i32, token_id: &String, conn: &SqliteConnection) {
+        use crate::database::schema::subscription::dsl::*;
+        let exists = select(exists(subscription.filter(
+            (restaurant_id.eq(res_id)).and(token.eq(token_id))
+        )
+        )).first::<bool>(conn);
+        match exists {
+            Ok(_) => {},
+            Err(_) => {
+                let new_sub = (
+                    restaurant_id.eq(res_id),
+                    token.eq(token_id),
+                );
+                insert_into(subscription).values(&new_sub).execute(conn).expect("Couldn't create subscription");
             }
         }
     }
 
-    pub fn find_favorite(res_id: i32, u_id: String, conn: &SqliteConnection) -> Option<Self> {
-        use super::schema::favorites::dsl::restaurant_id;
-        use super::schema::favorites::dsl::token_id;
-
-        fav_dsl
-            .filter(restaurant_id.eq(res_id))
-            .filter(token_id.eq(u_id))
-            .first::<Favorites>(conn)
-            .ok()
-    }
-    pub fn user_favorites(u_id: String, conn: &SqliteConnection) -> Vec<Self> {
-        use super::schema::favorites::dsl::token_id;
-        fav_dsl
-            .filter(token_id.eq(u_id))
-            .get_results::<Favorites>(conn)
-            .expect("Error loading posts")
-    }
-}
-
-use super::schema::users;
-use super::schema::users::dsl::users as u_dsl;
-
-#[derive(Debug, Deserialize, Serialize, Queryable, Insertable)]
-#[table_name = "users"]
-pub struct User {
-    pub token_id: String,
-    pub notifications: i32,
-}
-
-impl User {
-    pub fn new_user(id: String, conn: &SqliteConnection) {
-        let a_user = User {
-            token_id: id,
-            notifications: 0,
-        };
-        match u_dsl.find(&a_user.token_id).first::<User>(conn).ok() {
-            None => {
-                diesel::insert_into(u_dsl)
-                    .values(&a_user)
-                    .execute(conn)
-                    .expect("failed to insert new user");
-            }
-            Some(_) => {}
-        }
-    }
-
-    pub fn notification_change(id: String, conn: &SqliteConnection) {
-        use super::schema::users::dsl::notifications;
-        use super::schema::users::dsl::token_id;
-        let a_user = u_dsl.find(id).first::<User>(conn).ok();
-        match a_user {
-            None => {}
-            Some(x) => {
-                diesel::update(u_dsl.filter(token_id.eq(x.token_id)))
-                    .set(notifications.eq(if x.notifications == 1 { 0 } else { 1 }))
-                    .execute(conn)
-                    .expect("error updating notification setting");
-            }
-        }
-    }
-
-    pub fn remove_user(id: String, conn: &SqliteConnection) {
-        use super::schema::users::dsl::token_id;
-        diesel::delete(u_dsl.filter(token_id.eq(id)))
-            .execute(conn)
-            .expect("error removing user");
+    pub fn unsubscribe(res_id: i32, token_id: &String, conn: &SqliteConnection) {
+        use crate::database::schema::subscription::dsl::*;
+        delete(subscription.filter(
+            (restaurant_id.eq(res_id)).and(token.eq(token_id))
+        )).execute(conn).expect("Couldn't delete subscription");
     }
 }
