+++ src/services/mod.rs
@@ -1,0 +1,3 @@
+pub mod example;
+pub mod restaurant;
+pub mod subscription;
